Fix code location and add sqlite to gradle
