+++ app/build.gradle
@@ -24,6 +24,7 @@ dependencies {
     implementation 'com.google.guava:guava:31.1-jre'
     implementation 'com.formdev:flatlaf:3.0'
     implementation group: 'mysql', name: 'mysql-connector-java', version: '8.0.32'
+    implementation group: 'org.xerial', name: 'sqlite-jdbc', version: '3.40.0.0'
 }
 
 application {
