+++ app/src/main/java/clothing4you/JDBC.java
@@ -8,10 +8,18 @@ public class JDBC {
     public static String createUserTable = "CREATE TABLE users (" +
             "ID INTEGER PRIMARY KEY AUTOINCREMENT," +
             "first_name TEXT," +
-            "last_name TEXT," +
             "email TEXT," +
             "username TEXT," +
             "password TEXT" +
+            ");";
+
+    public static String createCatalogTable = "CREATE TABLE users (" +
+            "ID INTEGER PRIMARY KEY AUTOINCREMENT," +
+            "name TEXT," +
+            "category TEXT," +
+            "size TEXT," +
+            "quantity INTEGER," +
+            "price FLOAT(2)" +
             ");";
 
     public JDBC() throws SQLException, ClassNotFoundException {
@@ -26,7 +34,8 @@ public class JDBC {
             Class.forName("org.sqlite.JDBC");
             conn = DriverManager.getConnection("jdbc:sqlite:clothing4you.db");
 
-        } catch ( Exception e ) {
+        } catch (Exception e) {
+            throw e;
         }
         return conn;
     }
@@ -43,19 +52,32 @@ public class JDBC {
         System.out.println("Table created successfully");
     }
 
-   public static void insertUser(String first_name, String last_name, String email,String username, String password) throws SQLException, ClassNotFoundException {
+   public static void insertUser(String first_name, String email,String username, String password) throws SQLException, ClassNotFoundException {
         // Inserts user into user table to allow for data query
 
         Connection conn = establishConnection();
         conn.setAutoCommit(false);
         Statement stmt = conn.createStatement();
-        String sqlStatement = "INSERT INTO users (first_name,last_name,email,username,password) " +
-               "VALUES ('" + first_name + "','" + last_name + "','" + email + "','" +username +"','"+ password + "');";
+        String sqlStatement = "INSERT INTO users (first_name,email,username,password) " +
+               "VALUES ('" + first_name + "','" + email + "','" +username +"','"+ password + "');";
         stmt.executeUpdate(sqlStatement);
 
         conn.commit();
         conn.close();
    }
+    public static void insertItem(String name, String category, String size, int quantity, double price) throws SQLException, ClassNotFoundException {
+        // Inserts user into user table to allow for data query
+
+        Connection conn = establishConnection();
+        conn.setAutoCommit(false);
+        Statement stmt = conn.createStatement();
+        String sqlStatement = "INSERT INTO users (name,category,size,quantity,price) " +
+                "VALUES ('" + name + "','" + category + "','" + size + "','" +quantity +"','"+ price + "');";
+        stmt.executeUpdate(sqlStatement);
+
+        conn.commit();
+        conn.close();
+    }
 
     public static ArrayList query(String table, String column_name) throws SQLException, ClassNotFoundException {
         // Queries given table and returns array with data in column_name
@@ -66,7 +88,21 @@ public class JDBC {
         ResultSet resultSet = statement.executeQuery("SELECT * from " + table);
 
         while(resultSet.next()) {
-            result.add(resultSet.getString(column_name));
+            if(table == "catalog") {
+                String name = (resultSet.getString("name"));
+                String category = (resultSet.getString("category"));
+                String size = (resultSet.getString("size"));
+                String quantity = (resultSet.getString("quantity"));
+                String price = (resultSet.getString("price"));
+                result.add(name + " " + category + " " + size + " " + quantity + " " + price);
+            }
+            else {
+                String first_name = (resultSet.getString("name"));
+                String email = (resultSet.getString("category"));
+                String username = (resultSet.getString("size"));
+                String password = (resultSet.getString("quantity"));
+                result.add(first_name + " " + email + " " + username + " " + password);
+            }
             //System.out.println(resultSet.getString("email"));
         }
 
@@ -91,11 +127,11 @@ public class JDBC {
     }
     // For reference
     public static void main(String[] args) throws SQLException, ClassNotFoundException {
-//        createTable("users", createUserTable);
-//        insertUser("andrija", "rad", "a@gmail.com","andrija121", "1234");
-//        insertUser("willy", "lego", "willlego@gmail.com","willo1053", "1234");
-//        ArrayList a = query("users", "username");
-//        System.out.println(a);
-//        System.out.println(exists("wow14266", "users", "username"));
+        createTable("users", createUserTable);
+        insertUser("andrija", "a@gmail.com","andrija121", "1234");
+        insertUser("willy",  "willlego@gmail.com","willo1053", "1234");
+        ArrayList a = query("users", "username");
+        System.out.println(a);
+        System.out.println(exists("wow14266", "users", "username"));
     }
 }
