--- clothing4you/Login.form
@@ -1,95 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="clothing4you.Login">
-  <grid id="27dc6" binding="loginPanel" layout-manager="GridLayoutManager" row-count="5" column-count="4" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="20" y="20" width="604" height="624"/>
-    </constraints>
-    <properties>
-      <background color="-15592942"/>
-      <enabled value="false"/>
-    </properties>
-    <border type="none"/>
-    <children>
-      <vspacer id="7356c">
-        <constraints>
-          <grid row="1" column="3" row-span="2" col-span="1" vsize-policy="6" hsize-policy="1" anchor="0" fill="2" indent="0" use-parent-layout="false"/>
-        </constraints>
-      </vspacer>
-      <component id="4f219" class="javax.swing.JLabel">
-        <constraints>
-          <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="1" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-          <text value="Username:"/>
-        </properties>
-      </component>
-      <component id="19861" class="javax.swing.JTextField" binding="tfUsername">
-        <constraints>
-          <grid row="1" column="1" row-span="1" col-span="2" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="-1"/>
-          </grid>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-        </properties>
-      </component>
-      <component id="fa2ef" class="javax.swing.JButton" binding="btnLogin">
-        <constraints>
-          <grid row="3" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-          <text value="Login"/>
-        </properties>
-      </component>
-      <component id="e3100" class="javax.swing.JButton" binding="btnRegister">
-        <constraints>
-          <grid row="3" column="2" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-          <text value="Register"/>
-        </properties>
-      </component>
-      <component id="30100" class="javax.swing.JPasswordField" binding="pfPassword">
-        <constraints>
-          <grid row="2" column="1" row-span="1" col-span="2" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="-1"/>
-          </grid>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-        </properties>
-      </component>
-      <component id="73081" class="javax.swing.JLabel">
-        <constraints>
-          <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="1" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-          <labelFor value="30100"/>
-          <text value="Password:"/>
-        </properties>
-      </component>
-      <component id="fe89e" class="javax.swing.JButton" binding="btnGuest">
-        <constraints>
-          <grid row="4" column="1" row-span="1" col-span="2" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-          <text value="Continue As Guest"/>
-        </properties>
-      </component>
-      <component id="c4627" class="javax.swing.JLabel" binding="iconLabel">
-        <constraints>
-          <grid row="0" column="0" row-span="1" col-span="3" vsize-policy="0" hsize-policy="0" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value="Label"/>
-        </properties>
-      </component>
-    </children>
-  </grid>
-</form>
