--- clothing4you/Login.java
@@ -1,75 +1,0 @@
-package clothing4you;
-
-
-import javax.swing.*;
-import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.sql.SQLException;
-
-
-public class Login extends JDialog {
-
-    private JPanel loginPanel;
-    private JTextField tfUsername;
-    private JButton btnLogin;
-    private JButton btnRegister;
-    private JPasswordField pfPassword;
-    private JButton btnGuest;
-    private JLabel iconLabel;
-
-    ImageIcon icon = new ImageIcon("img/Icon.png");
-
-    public Login(JFrame parent) {
-        super(parent);
-        setTitle("Login");
-        setContentPane(loginPanel);
-        setMinimumSize(new Dimension(600, 600));
-        setDefaultCloseOperation(WindowConstants.DISPOSE_ON_CLOSE);
-        setModal(true);
-        setLocationRelativeTo(parent);
-
-        iconLabel.setText("");
-        iconLabel.setIcon(icon);
-
-        btnLogin.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                String username = tfUsername.getText();
-                char[] password = pfPassword.getPassword();
-                dispose();
-                try {
-                    Catalog myCatalog = new Catalog(null);
-                } catch (SQLException ex) {
-                    throw new RuntimeException(ex);
-                } catch (ClassNotFoundException ex) {
-                    throw new RuntimeException(ex);
-                }
-            }
-        });
-        btnRegister.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                dispose();
-                Register myRegister = new Register(null);
-            }
-        });
-        btnGuest.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                dispose();
-                try {
-                    Catalog myCatalog = new Catalog(null);
-                } catch (SQLException ex) {
-                    throw new RuntimeException(ex);
-                } catch (ClassNotFoundException ex) {
-                    throw new RuntimeException(ex);
-                }
-            }
-        });
-
-        setVisible(true);
-    }
-
-
-}
