--- clothing4you/Return.form
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="clothing4you.Return">
-  <grid id="27dc6" binding="returnPanel" layout-manager="GridLayoutManager" row-count="4" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="20" y="20" width="500" height="400"/>
-    </constraints>
-    <properties/>
-    <border type="none"/>
-    <children>
-      <component id="350ff" class="javax.swing.JLabel">
-        <constraints>
-          <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-          <text value="Product Name:"/>
-        </properties>
-      </component>
-      <component id="2b32c" class="javax.swing.JLabel">
-        <constraints>
-          <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-          <text value="Reason for Return:"/>
-        </properties>
-      </component>
-      <component id="7e3fd" class="javax.swing.JLabel">
-        <constraints>
-          <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-          <text value="Quantity:"/>
-        </properties>
-      </component>
-      <component id="dc993" class="javax.swing.JSpinner" binding="quantitySpinner">
-        <constraints>
-          <grid row="2" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-        </properties>
-      </component>
-      <component id="caa58" class="javax.swing.JTextField" binding="returnTF">
-        <constraints>
-          <grid row="1" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="-1"/>
-          </grid>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-        </properties>
-      </component>
-      <component id="bc6b9" class="javax.swing.JTextField" binding="nameTF">
-        <constraints>
-          <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="-1"/>
-          </grid>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-        </properties>
-      </component>
-      <component id="a8eb5" class="javax.swing.JButton" binding="submitBtn">
-        <constraints>
-          <grid row="3" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-          <text value="Submit"/>
-        </properties>
-      </component>
-      <component id="f6b2d" class="javax.swing.JButton" binding="cancelBtn">
-        <constraints>
-          <grid row="3" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="Times New Roman" size="16" style="1"/>
-          <text value="Cancel"/>
-        </properties>
-      </component>
-    </children>
-  </grid>
-</form>
