--- clothing4you/User.java
@@ -1,47 +1,0 @@
-package clothing4you;
-
-public class User {
-    private String name;
-    private String username;
-    private String email;
-    private char[] password;
-
-    public User(String name, String username, String email, char[] password) {
-        this.name = name;
-        this.username = username;
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public char[] getPassword() {
-        return password;
-    }
-
-    public void setPassword(char[] password) {
-        this.password = password;
-    }
-}
